+++ lib/src/util/asset_util.dart
@@ -1,0 +1,5 @@
+class AssetUtil {
+  static bool isIcon(String assetName) =>
+      assetName.startsWith('assets/icons/') ||
+      assetName.startsWith('assets/images/icons/');
+}